cgo: mark some typedefs not as Go type aliases
To avoid horrible diagnostic messages on type errors, use type
declarations that are not aliases for some typedefs. A common case is
code like the following:

    typedef struct {
        int foo;
        int bar;
    } foobar_t;

This struct would be dumped inline in error messages, which is
especially problematic for large structs. To avoid that, do not alias
these.

This has the potential to break some code, but the current tests still
pass. The heuristc should be improved if there is code that is broken by
this change.